moving api-user to the api folder
