+++ client/containers/Users.js
@@ -12,7 +12,7 @@ import Typography from 'material-ui/Typography';
 import ArrowForward from 'material-ui-icons/ArrowForward';
 import Person from 'material-ui-icons/Person';
 import { Link } from 'react-router-dom';
-import { list } from './api-user.js';
+import { list } from '../api/api-user.js';
 
 const styles = theme => ({
   root: theme.mixins.gutters({
